Separate library from test properties
